third_party/blink/web_tests/http/tests/security/mixedContent/insecure-frame-in-data-iframe-in-main-frame-blocked.html

<html>
<body>
<script>
if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
    testRunner.setPopupBlockingEnabled(false);
    testRunner.overridePreference("WebKitAllowRunningInsecureContent", false);
}

window.addEventListener("message", function (e) {
  if (window.testRunner)
      testRunner.notifyDone();
}, false);

</script>
<p>This test opens a window that loads a data: iframe that loads an
insecure frame, and that the frame is still blocked.</p>
<script>
window.open("https://127.0.0.1:8443/security/mixedContent/resources/frame-with-data-url-frame-with-frame.html");
</script>
</body>
</html>
